# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: prompts
    description: Prompt operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /prompts:
    get:
      operationId: listPrompts
      tags: [prompts]
      summary: List all prompts
      description: Returns all prompts ordered by creation date descending
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromptSummary"
    post:
      operationId: createPrompt
      tags: [prompts]
      summary: Create a new prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptBody"
      responses:
        "201":
          description: Created prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prompt"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /prompts/{id}:
    get:
      operationId: getPrompt
      tags: [prompts]
      summary: Get a single prompt
      description: Returns a single prompt and increments its view count
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Prompt with view count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptDetail"
        "404":
          description: Prompt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /prompts/stats:
    get:
      operationId: getPromptsStats
      tags: [prompts]
      summary: Get library stats
      description: Returns aggregate stats for the prompt library
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptsStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    PromptSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        complexity:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - complexity
        - createdAt

    Prompt:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        complexity:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - complexity
        - createdAt

    PromptDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        complexity:
          type: integer
        createdAt:
          type: string
          format: date-time
        viewCount:
          type: integer
      required:
        - id
        - title
        - content
        - complexity
        - createdAt
        - viewCount

    CreatePromptBody:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        content:
          type: string
          minLength: 20
        complexity:
          type: integer
          minimum: 1
          maximum: 10
      required:
        - title
        - content
        - complexity

    PromptsStats:
      type: object
      properties:
        total:
          type: integer
        avgComplexity:
          type: number
        complexityBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ComplexityCount"
        recentCount:
          type: integer
      required:
        - total
        - avgComplexity
        - complexityBreakdown
        - recentCount

    ComplexityCount:
      type: object
      properties:
        level:
          type: string
        count:
          type: integer
      required:
        - level
        - count

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error