# lib/api-zod/src/generated/types/complexityCount.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const ComplexityCount = zod.object({
  level: zod.string(),
  count: zod.number(),
});

export type ComplexityCount = zod.input<typeof ComplexityCount>;