# lib/api-zod/src/generated/types/createPromptBody.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const createPromptBodyTitleMinTwo = 3;

export const createPromptBodyContentMinTwo = 20;

export const createPromptBodyComplexityMaxTwo = 10;

export const CreatePromptBody = zod.object({
  title: zod.string().min(createPromptBodyTitleMinTwo),
  content: zod.string().min(createPromptBodyContentMinTwo),
  complexity: zod.number().min(1).max(createPromptBodyComplexityMaxTwo),
});

export type CreatePromptBody = zod.input<typeof CreatePromptBody>;