# lib/api-zod/src/generated/types/prompt.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const Prompt = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  complexity: zod.number(),
  createdAt: zod.coerce.date(),
});

export type Prompt = zod.input<typeof Prompt>;