# lib/api-zod/src/generated/types/promptDetail.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const PromptDetail = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  complexity: zod.number(),
  createdAt: zod.coerce.date(),
  viewCount: zod.number(),
});

export type PromptDetail = zod.input<typeof PromptDetail>;