# lib/api-zod/src/generated/types/promptSummary.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const PromptSummary = zod.object({
  id: zod.number(),
  title: zod.string(),
  complexity: zod.number(),
  createdAt: zod.coerce.date(),
});

export type PromptSummary = zod.input<typeof PromptSummary>;