# lib/api-zod/src/generated/types/promptsStats.zod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import { z as zod } from "zod";

export const PromptsStats = zod.object({
  total: zod.number(),
  avgComplexity: zod.number(),
  complexityBreakdown: zod.array(
    zod.object({
      level: zod.string(),
      count: zod.number(),
    }),
  ),
  recentCount: zod.number(),
});

export type PromptsStats = zod.input<typeof PromptsStats>;